+++ tests/testthat/test-geom_casting.R
@@ -26,3 +26,18 @@ test_that("available sets & shapes", {
   )
 })
 
+test_that("shapes clipped when zooming", {
+  file <- tempfile(fileext = ".png")
+
+  p <- ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) +
+    geom_casting(size = 0.1, shape = "violin") +
+    ggplot2::geom_point() +
+    ggplot2::coord_cartesian(xlim = c(1, 4))
+
+  ggplot2::ggsave(file, p, "png")
+
+  expect_snapshot_file(file)
+
+  unlink(file)
+})
+